Fixed build errors by updating dependencies
Nodejs buildpack was failing with Error /usr/bin/jq Permission denied.
I found this issue https://github.com/heroku/heroku-buildpack-nodejs/issues/856
The solution is to pin the buildpack version to an earlier version such as 175. This should probably be removed some time in the future.

PHP didn't want to build because ext-mbstring was missing. I added the package to composer.json but it still didn't work. I found this issue https://github.com/gliderlabs/herokuish/issues/529
This is broken somehow and the temporary solution is to manually apt install some packages using another buildpack. This should also probably be removed in the future.